pdfs-eval/requirements.txt
torch==2.6.0
torchvision==0.21.0
datasets==3.2.0
transformers==4.49.0
openai==1.61.1
qwen-vl-utils==0.0.10
accelerate==1.4.0
google-genai==1.0.0
easyocr==1.7.2
pytesseract==0.3.13
sacrebleu==2.5.1
pdf2image==1.17.0
marker-pdf==1.4.0
docling==2.21.0